+++ CakeMail.RestClient/Models/ListMember.cs
@@ -16,7 +16,7 @@ namespace CakeMail.RestClient.Models
 		/// The id.
 		/// </value>
 		[JsonProperty("id")]
-		public int Id { get; set; }
+		public long Id { get; set; }
 
 		/// <summary>
 		/// Gets or sets the status.
@@ -43,7 +43,7 @@ namespace CakeMail.RestClient.Models
 		/// The bounce_count.
 		/// </value>
 		[JsonProperty("bounce_count")]
-		public int BounceCount { get; set; }
+		public long BounceCount { get; set; }
 
 		/// <summary>
 		/// Gets or sets the email.
